Added support for skipping not found files
Added support for skipping not found files
